+++ src/se/nicklasgavelin/sphero/macro/command/Emit.java
@@ -1,6 +1,7 @@
 package se.nicklasgavelin.sphero.macro.command;
 
 import se.nicklasgavelin.sphero.macro.MacroCommand;
+import se.nicklasgavelin.sphero.macro.MacroCommand.MACRO_COMMAND;
 import se.nicklasgavelin.util.ByteArrayBuffer;
 import se.nicklasgavelin.util.Value;
 
@@ -13,7 +14,7 @@ public class Emit extends MacroCommand
 {
 	private static final int DEFAULT_IDENTIFIER = 1, MAX_IDENTIFIER = 255,
 			MIN_IDENTIFIER = 0;
-	private Integer identifier = Integer.valueOf( DEFAULT_IDENTIFIER );
+	private Integer identifier = DEFAULT_IDENTIFIER;
 
 	// public Emit( byte[] data )
 	// {
@@ -21,7 +22,7 @@ public class Emit extends MacroCommand
 	public Emit( int _identifier )
 	{
 		super( MACRO_COMMAND.MAC_EMIT );
-		this.setIdentifier( _identifier );
+        this.setIdentifier( _identifier );
 	}
 
 	/**
@@ -41,7 +42,7 @@ public class Emit extends MacroCommand
 	 */
 	public void setIdentifier( Integer _identifier )
 	{
-		this.identifier = Value.clamp( _identifier, MIN_IDENTIFIER, MAX_IDENTIFIER );
+        this.identifier = Value.clamp( _identifier, MIN_IDENTIFIER, MAX_IDENTIFIER);
 		// if ( (_identifier.intValue() >= IDENTIFIER_MIN_VALUE) &&
 		// (_identifier.intValue() <= IDENTIFIER_MAX_VALUE) )
 		// this.identifier = _identifier;
@@ -50,9 +51,9 @@ public class Emit extends MacroCommand
 	@Override
 	public byte[] getByteRepresentation()
 	{
-		ByteArrayBuffer bytes = new ByteArrayBuffer( getLength() );
-		bytes.append( getCommandID() );
-		bytes.append( this.identifier.intValue() );
+		ByteArrayBuffer bytes = new ByteArrayBuffer(getLength() );
+		bytes.append(getCommandID() );
+		bytes.append(this.identifier);
 
 		return bytes.toByteArray();
 	}
